+++ leia-http-processor/src/main/java/com/grookage/leia/http/processor/executor/HttpMessageExecutorFactory.java
@@ -20,6 +20,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.base.Preconditions;
 import com.grookage.leia.http.processor.config.HttpClientConfig;
 import com.grookage.leia.http.processor.endpoint.EndPointResolver;
+import com.grookage.leia.mux.executor.MessageExecutor;
+import com.grookage.leia.mux.executor.MessageExecutorFactory;
 
 import java.util.Locale;
 import java.util.Map;
@@ -27,10 +29,10 @@ import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Supplier;
 
-public class DefaultHttpExecutorFactory implements HttpExecutorFactory {
-    private final Map<String, HttpExecutor> executors = new ConcurrentHashMap<>();
+public class HttpMessageExecutorFactory implements MessageExecutorFactory {
+    private final Map<String, MessageExecutor> executors = new ConcurrentHashMap<>();
 
-    public DefaultHttpExecutorFactory(final HttpClientConfig clientConfig,
+    public HttpMessageExecutorFactory(final HttpClientConfig clientConfig,
                                       final Supplier<String> authSupplier,
                                       final EndPointResolver endPointResolver,
                                       final ObjectMapper mapper) {
@@ -39,7 +41,7 @@ public class DefaultHttpExecutorFactory implements HttpExecutorFactory {
         Preconditions.checkNotNull(mapper, "Object mapper can't be null");
         Preconditions.checkNotNull(endPointResolver, "Endpoint Resolver can't be null");
         clientConfig.getBackendConfigs()
-                .forEach(backendConfig -> executors.put(backendConfig.getBackendName().toUpperCase(Locale.ROOT), new DefaultHttpExecutor(
+                .forEach(backendConfig -> executors.put(backendConfig.getBackendName().toUpperCase(Locale.ROOT), new HttpMessageExecutor(
                         backendConfig,
                         authSupplier,
                         mapper,
@@ -48,7 +50,7 @@ public class DefaultHttpExecutorFactory implements HttpExecutorFactory {
     }
 
     @Override
-    public Optional<HttpExecutor> getExecutor(String backendName) {
+    public Optional<MessageExecutor> getExecutor(String backendName) {
         return Optional.ofNullable(executors.get(backendName.toUpperCase(Locale.ROOT)));
     }
 }