+++ leia-processor/src/main/java/com/grookage/leia/mux/targetvalidator/JsonRuleTargetValidator.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.grookage.leia.mux.processors;
+package com.grookage.leia.mux.targetvalidator;
 
 import com.grookage.leia.models.mux.MessageRequest;
 import com.grookage.leia.models.schema.SchemaDetails;
@@ -22,12 +22,15 @@ import com.grookage.leia.models.schema.transformer.TransformationTarget;
 import lombok.AllArgsConstructor;
 
 @AllArgsConstructor
-public class DefaultTargetValidator implements TargetValidator {
+public class JsonRuleTargetValidator implements TargetValidator {
 
     @Override
     public boolean validate(TransformationTarget transformationTarget,
                             MessageRequest messageRequest,
                             SchemaDetails schemaDetails) {
-        return true;
+        if (null == transformationTarget.getCriteria()) { // If there is null, treat this NO_CRITERIA, return true!
+            return true;
+        }
+        return transformationTarget.getCriteria().evaluate(messageRequest.getMessage());
     }
 }