+++ leia-processor/src/test/java/com/grookage/leia/mux/resolver/TagBasedNameResolverTest.java
@@ -14,23 +14,23 @@
  * limitations under the License.
  */
 
-package com.grookage.leia.http.processor.resolver;
+package com.grookage.leia.mux.resolver;
 
-import com.grookage.leia.http.processor.config.HttpClientConfig;
 import com.grookage.leia.models.ResourceHelper;
 import com.grookage.leia.models.mux.LeiaMessage;
+import com.grookage.leia.mux.resolver.TagBasedNameResolver;
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-class TagBasedBackendNameResolverTest {
+import java.util.List;
+
+class TagBasedNameResolverTest {
 
     @Test
     @SneakyThrows
     void testNameResolver() {
-        final var httpClientConfig = ResourceHelper.getResource("httpClientConfig.json", HttpClientConfig.class);
-        Assertions.assertNotNull(httpClientConfig);
-        final var resolver = new TagBasedBackedNameResolver(httpClientConfig);
+        final var resolver = new TagBasedNameResolver(() -> List.of("BACKEND1"));
         final var leiaMessage = ResourceHelper.getResource("mux/leiaMessage.json", LeiaMessage.class);
         Assertions.assertNotNull(leiaMessage);
         final var eligibleBackends = resolver.getEligibleBackends(leiaMessage);